add stories section
This commit adds a ‘stories’ section to the website, accessible
from the ‘Stories’ link in the top right of the menu.

You can create new posts by adding a markdown file in
`/content/stories/`.

The file requires some frontmatter parameters:

```
---
url: /blog/hello-world
title: "Hello world"
author: "Jane Dodds"
date: 2020-11-17
---
```

As with markdown files, if you also want to show a table of contents
in this side bar, you can add `show_table_of_contents: true`, which
will render out a list of links to all H2 and H3 elements in the
markdown file.

You can chose to omit the date, or author and they won’t be printed.